requirements.txt
feedparser==6.0.11
groq>=0.9.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
